Update Core System #2
- add more products, ingredients, catalog to tidy up editor view.
-Improve SearchDrink logic where if the stall doesn’t have the signature item or product, it will check for other products that it fulfill it requirement, if fail then buy the minimum product that it needs like bottle of water, if it still fail then leave the stall.

- now CoreController keep track of PersonView gameobject instead of Person class.
- the logic that handle NPC(Person.cs) will live inside PersonView
- PersonView will provide public function to update the logic

+++ Assets/ProjectSims/Simulation/CoreSystem/InventoryController.cs
@@ -9,14 +9,14 @@ namespace Simulation.Inventory
 {
     public class InventoryController
     {
-        private Dictionary<Item, int> _dictInventory;
+        private Dictionary<Item, float> _dictInventory;
 
         public InventoryController()
         {
-            _dictInventory = new Dictionary<Item, int>(8);
+            _dictInventory = new Dictionary<Item, float>(8);
         }
 
-        public void Add(Item item, int qty)
+        public void Add(Item item, float qty)
         {
             var boolean = _dictInventory.TryAdd(item, qty);
             if (!boolean)
@@ -25,7 +25,7 @@ namespace Simulation.Inventory
             }
         }
 
-        public bool Get(Item item, int amount)
+        public bool Get(Item item, float amount)
         {
             if (!_dictInventory.ContainsKey(item))
             {
@@ -42,7 +42,7 @@ namespace Simulation.Inventory
             return true;
         }
 
-        public bool CheckItem(Item item, int amount)
+        public bool CheckItem(Item item, float amount)
         {
             if (!_dictInventory.ContainsKey(item))
             {
